# packages/backend/app/openapi.yaml
openapi: 3.0.3
info:
  title: FinMind API
  version: 0.1.0
  description: API for budgeting, bills, reminders and insights
servers:
  - url: http://localhost:8000
...
paths:
...
  /reminders:
    get:
      summary: List reminders
      tags: [Reminders]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
              example:
                - { id: 6, message: Pay credit card, send_at: 2025-08-11T10:00:00Z, sent: false, channel: email }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    post:
      summary: Create reminder
      tags: [Reminders]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewReminder'
            example:
              message: Pay credit card
              send_at: 2025-08-11T10:00:00Z
              channel: email
      responses:
        '201': { description: Created }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /reminders/run:
    post:
      summary: Process due reminders
      tags: [Reminders]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Processed count }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /reminders/bills/{billId}/schedule:
    post:
      summary: Create bill reminders with default/override offsets
      tags: [Reminders]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: billId
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                offsets_days:
                  type: array
                  items: { type: integer }
      responses:
        '200':
          description: Created reminders count
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: integer }
  /reminders/bills/{billId}/autopay-result:
    post:
      summary: Emit autopay outcome follow-up reminders
      tags: [Reminders]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: billId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [SUCCESS, FAILED] }
      responses:
        '200':
          description: Created reminders count
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: integer }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      example: { error: "invalid credentials" }
...
    Reminder:
      type: object
      properties:
        id: { type: integer }
        message: { type: string }
        send_at: { type: string, format: date-time }
        sent: { type: boolean }
        channel: { type: string, enum: [email, whatsapp] }
    NewReminder:
      type: object
      required: [message, send_at]
      properties:
        message: { type: string }
        send_at: { type: string, format: date-time }
        channel: { type: string, enum: [email, whatsapp], default: email }